requirements.txt
pymongo==4.6.0
faker==20.1.0
pandas==2.1.1
numpy==1.26.0
matplotlib==3.8.0
seaborn==0.13.0
psutil==5.9.6
pytest==7.4.3
python-dotenv==1.0.0
tqdm==4.66.1